move handlers directory under structures directory
