+++ src/lib/structures/handlers/event.handler.ts
@@ -1,9 +1,9 @@
 import { Client, ClientEvents } from 'discord.js';
 import { redBright, yellowBright } from 'colorette';
 import Handler from './handler';
-import Event from '../structures/pieces/event.piece';
-import EventStore from '../structures/stores/event.store';
-import Container from '../container';
+import Event from '../pieces/event.piece';
+import EventStore from '../stores/event.store';
+import Container from '../../container';
 
 class EventHandler implements Handler {
   private eventStore: EventStore;
