+++ src/lib/structures/handlers/handler.ts
@@ -1,5 +1,5 @@
 import { Client } from 'discord.js';
-import Container from '../container';
+import Container from '../../container';
 
 interface Handler {
   initialize(client: Client, container: Container): Promise<void>;
